Refactor event emitter tests directory
