+++ tests/__init__.py
@@ -1,4 +1,1 @@
 from .conftest import EventFixtures, CallbackFixtures, CeleryMock, CelerySerializerMock
-from .emitters import *
-from .handlers import *
-from .linkers import *
