--- tests/emitters/__init__.py
@@ -1,4 +1,0 @@
-from .event_emitter_test import EventEmitterTest
-from .test_asyncio_event_emitter import TestAsyncIOEventEmitter
-from .test_executor_event_emitter import TestExecutorEventEmitter
-from .test_rq_event_emitter import TestRQEventEmitter
