--- tests/events/__init__.py
@@ -1,1 +1,0 @@
-from .test_event import TestEvent
