--- tests/handlers/__init__.py
@@ -1,1 +1,0 @@
-from .test_event_handler import TestEventHandler
