--- tests/linkers/__init__.py
@@ -1,1 +1,0 @@
-from .test_event_linker import TestEventLinker
